requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.40 